# Optimize speed and memory usage on large text files (>1 megabyte)