(Ok, the title might sound funny because files of a few megabytes in size are considered large in 2021 when the RAM amounts of our computers are measured in gigabytes, but still, that's no reason not to optimize...)

This PR is pretty long, I know, but it contains a few intertwined optimizations, so I thought it would make sense to apply them all together. Here is a list of all stuff that I have done:

1. So first of all, I applied the general LuaJIT recommendations, such as avoiding `table.insert` for appending to lists and instead using a counter, not calculating the lengths of tables and storing their lengths instead, and, of course, recycling and clearing existing tables instead of allocating new ones as often as possible.

2. Instead of storing a table of compiled regexes in each instance of Buffer, I just compile the regex once in the constructor. AFAIK compilation should be independent of the editor state (with the exception of the `regexpengine` option, definitely), unlike evaluation, which is why I added `buf_call`s in my previous PR. Because a cache of regexes was stored separately in each Buffer, it would only contain a single item, which would have to be created almost immediately after construction in the `index_line` method anyway.

3. Speaking of `index_line`, I didn't record a proper profile, but I feel like evaluation of regexes takes roughly 90% of the whole time spent on indexing. This is really hard to optimize because most of the work is spent in Vim's regex engine and not in Lua where we have control. One cool optimization I imagined is offloading indexing to an entire separate thread with `vim.loop.new_work`, but that won't work because a) I doubt that Vim's regex engine is thread-safe and so this would explode and segfault; b) we need to call `nvim_buf_call`, which is definitely not thread-safe. I can imagine just re-implementing some simple specific regex with Lua code, for instance, `\k\+` (or adding a special case for this regex), which would require re-implementing `iskeyword` (the most interesting parts are [`buf_init_chartab`](https://github.com/neovim/neovim/blob/v0.5.0/src/nvim/charset.c#L83-L262) and [`vim_iswordc_tab`](https://github.com/neovim/neovim/blob/v0.5.0/src/nvim/charset.c#L831-L843) in `charset.c`, and the handling of non-ASCII characters ([`utf_class_tab`](https://github.com/neovim/neovim/blob/v0.5.0/src/nvim/mbyte.c#L1069-L1182) in `mbyte.c`, yeah...), and, in fact, this is [what coc.nvim does](https://github.com/neoclide/coc.nvim/blob/03c9add7cd867a013102dcb45fb4e75304d227d7/src/model/chars.ts) (although their implementation is incomplete, e.g. it doesn't try to do any special handling of Unicode characters and just considers all of them word chars).

   However, I believe that the ability to set a custom regex for recognizing words is a very convenient feature of this plugin, so let's leave it as it is. Anyway, there is still one way of optimizing `index_line` left - simply reduce the number of regex invocations. I don't know if this was intended, but the previous implementation had an inefficiency in it: when no match is found, it would advance character-by-character and retry the regex every time, instead of just skipping further processing of the line. Fixing this helped a little bit in reducing the run time (). Also, the previous implementation would skip 1 character after the match, which resulted in loss of some words (I have actually tested this by dumping the lists of indexed words).

4. Alright, now, for the main optimization related to RAM usage. I've added a "synchronous" indexing mode which takes advantage of some weird garbage collector behavior, and so **it improves memory usage by a factor of 2!!!** (I'll show results later) You may be worried by the fact that it is synchronous, however, on my machine at least it can process a 9 megabyte file in 1.5 seconds, so I believe that memory usage reduction is worth it. The trick that I use here is requesting buffer contents with multiple consequent `nvim_buf_get_lines` requests. I measured the overhead of making many many API calls here, in our case it is negligible in comparison to the time taken matching the regexes, so it's fine. I have also added options (with documentation) for controlling the speed of indexing and switching between the sync and async algorithms.

5. The deduplicated words are also cached. They are not updated right away in the watch callback, but rather on demand, when the `get_words` method sees that the lines have been modified by checking a "dirty" flag, because the `on_lines` event is invoked VERY frequently, and we don't want to block it with useless repeated work. Deduplication is not that fast, tens of milliseconds on 5-10 megabyte files, but the thing is that nvim-cmp calls the `complete` method of the source only once (on the first keystroke), so by itself this optimization will only save time if you, for example, trigger the popup menu, then close it, then open it again.

   However, the technique can be used in other ways, for instance: because completion is used only in the Insert mode and 99% of Insert editing in Vim happens on just the current line, the unique word table can be split in three parts: the current line, lines before it and lines after it, each with its own "dirty" flag. Then, the `on_lines` event will check if `first_line == current_line and first_line == new_last_line + 1 and old_last_line == new_last_line`, which would mean that only the current line has been changed, and only set the flag of the current line words table in that case. In general, this whole optimization is still effective even considering that we have to deduplicate words across multiple buffers (when the `get_bufnrs` option returns multiple IDs) because at this stage we are looping through hundreds of thousands of words.

6. Last but not least: I have rewritten the loops which append and delete lines in the `on_lines` event listener. See, the way it was written with Lua's built-in functions, which can only delete or insert one element at a time, meant that those loops had effectively O(n^2) complexity, which could very trivially result in huge lags when, for example, deleting the whole buffer contents with `ggdG`. You would think that clearing out the buffer would happen instantaneously because we only need to unload a bunch of strings from memory and there is nothing left to index, but no, the previous implementation would hang because of the quadratic time usage. I rewrote them to delete or move all elements in one loop, so this should not be a problem anymore. And also, I added the missing event handlers for `on_reload` (an easy way of triggering a reload: `:!touch %`) and `on_detach`, the last one is necessary to unload the indexes of closed buffers, so it kind of counts as a RAM optimization as well?

And now, the results. I'm going to use a compiled file from the Typescript repository, <https://github.com/microsoft/TypeScript/blob/v4.4.3/lib/tsserver.js>. It is perfect for the test because it is 10 megabytes big, is not minified (you will see why it is important later), and can be easily downloaded. However, for the purposes of the benchmark, I apply the following patch to the original code of this source:

```diff
diff --git a/lua/cmp_buffer/buffer.lua b/lua/cmp_buffer/buffer.lua
index 4ba7149..9fa86ef 100644
--- a/lua/cmp_buffer/buffer.lua
+++ b/lua/cmp_buffer/buffer.lua
@@ -42,7 +42,7 @@ function buffer.index(self)
     0,
     200,
     vim.schedule_wrap(function()
-      local chunk = math.min(index + 1000, #lines)
+      local chunk = #lines
       vim.api.nvim_buf_call(self.bufnr, function()
         for i = index, chunk do
           self:index_line(i, lines[i] or '')
```

Yes, it will turn the original asynchronous algorithm into a synchronous one because all will be lines indexed in one loop, but I think this doesn't matter and the comparison is still fair because we are not tricking the garbage collector here by running `nvim_buf_get_lines` multiple times, so how the memory is allocated won't change, and we will skip the pauses between indexing steps, so the CPU will be 100% loaded. I am going to measure performance with:

```lua
function _G._nvim_cmp_buffer_benchmark()
  local cmp_core = require('cmp.core')
  local source = cmp_core.sources_by_name['buffer'][1]
  local start_mem = vim.loop.resident_set_memory()
  local start_time = vim.loop.hrtime()
  source.source:complete({
    option = source:get_config().opts,
    context = { cursor_before_line = '' },
    offset = 1,
  }, function(result)
    local end_time = vim.loop.hrtime()
    local end_mem = vim.loop.resident_set_memory()
    print(
      string.format(
        '%d items, %f ms, %f MB',
        #result.items, -- For proof that everything works as before
        (end_time - start_time) / 1e6,
        (end_mem - start_mem) / 1024 / 1024
      )
    )
  end)
end
vim.cmd([[ autocmd VimEnter * lua vim.schedule(_G._nvim_cmp_buffer_benchmark) ]])
```

Let's go! The original version (5dde5430757696be4169ad409210cf5088554ed6):

```
run #1: 36667 items, 2069.154921 ms, 70.953125 MB
run #2: 36667 items, 2071.531198 ms, 70.949219 MB
run #3: 36667 items, 2047.786888 ms, 73.046875 MB
run #4: 36667 items, 2124.546513 ms, 70.917969 MB
run #5: 36667 items, 2058.463669 ms, 73.035156 MB
   avg:              2074.296638 ms, 71.780469 MB
```

Optimized version (b1fa2cb1dd71af0f072b7d2fa73ade3a7efdf01d), async mode (I am setting `indexing_chunk_size = 1e10` to get the same behavior as the patch shown above, therefore `indexing_interval` doesn't matter):

```
run #1: 36667 items, 1417.362432 ms, 68.656250 MB
run #2: 36667 items, 1420.224233 ms, 69.257812 MB
run #3: 36667 items, 1393.939492 ms, 68.683594 MB
run #4: 36667 items, 1402.539505 ms, 68.679688 MB
run #5: 36667 items, 1429.598909 ms, 69.207031 MB
   avg:              1412.732914 ms, 68.896875 MB
```

Optimized version (b1fa2cb1dd71af0f072b7d2fa73ade3a7efdf01d), sync mode (`indexing_interval = -1`):

```
run #1: 36667 items, 1424.161060 ms, 38.667969 MB
run #2: 36667 items, 1439.539344 ms, 35.007812 MB
run #3: 36667 items, 1448.200662 ms, 34.949219 MB
run #4: 36667 items, 1411.546087 ms, 38.914062 MB
run #5: 36667 items, 1406.199709 ms, 39.417969 MB
   avg:              1425.929372 ms, 37.391406 MB
```

Great! However, there are still two major performance problems left. First of all, `index_line` blows up on files with long lines. Why? You see, when we slice a string with `string.sub`, this function creates a copy of the slice, and this results in catastrophic memory usage on files with very long lines where very very long slices of the remaining part of the string are created after each match while running `index_line`, which just crashes Vim with a "not enough memory" error on my computer. The Vimscript `matchstrpos` function avoids this problem - it has a `start` parameter, so that we can start at any index we want without having to create and allocate a slice of the string starting at that index. But it is not available in Lua's `vim.regex.match_str`! This can only be solved by making a patch to Neovim: the [internal C function](https://github.com/neovim/neovim/blob/v0.5.0/src/nvim/regexp.c#L7387-L7392) ([documentation here](https://github.com/neovim/neovim/blob/v0.5.0/src/nvim/regexp.c#L7308-L7321)) for evaluating a regex actually has a parameter for the byte position at which to start searching, they just forgot to add to in the [Lua interface](https://github.com/neovim/neovim/blob/v0.5.0/src/nvim/lua/executor.c#L1601). Alternatively, we can just call the Vimscript function, which makes the indexing considerably slower (2517.396650 ms and 43.835938 MB on the file in the benchmarks) and recompiles the regex every time:

```lua
function buffer.index_line(self, linenr, line)
  -- ...
  local idx = 0
  local line_len = #line
  local matchstrpos = vim.fn.matchstrpos
  while idx < line_len do
    local word, match_start, match_end = unpack(matchstrpos(line, self.pattern, idx), 1, 3)
    if match_start >= 0 and match_end >= 0 then
      if #word > 1 then
        words[word_i] = word
        word_i = word_i + 1
      end
      idx = match_end
    else
      break
    end
  end
end
```

The other issue is, well, that nvim-cmp doesn't like when tens of thousands of unique words are being thrown at it. Again, taking the file from the benchmark, if I just open the editor with `tsserver.js`, it will take 35 megabytes of RAM, if I then run the indexer, it will consume 35 more, but just pressing `<Ctrl-Space>`, which triggers the completion menu, makes Vim eat 180 megabytes! I believe it is because of the need to create 37 thousand `cmp.Entry`ies, and I can't really blame nvim-cmp for that. The number of completion items definitely must be reduced, but I don't know what is the best way to do that. I can come up with two solutions (they can be implemented together):

1. An option like `word_min_occurrences` which would allow filtering out rare words which occur in the file just 1-2 times (but will probably make the user experience worse on typical, "small", code files).
2. Another option like `max_top_words` which would only show, for example, the top 1000 words in the completion results sorted by the number of their usages.

Both should be easy to do - we already have a table of unique words, and instead of setting `true` we can use each value as a counter. I addressed the problem of _indexing_ large files, but I believe that finding an answer to this particular question (reduction of the number of completion items) will help with improving the user experience of _editing_ large files.