@hrsh7th 2021-09-19
I've received the issue report (via twitter) that will make freeze the nvim when editing X mb buffer.

So I think we should introduce `max_file_bytes` or `line_count` option.