@hrsh7th 2021-09-19
It's hard to review. So it will be take a time.

Hm... Probably, we should remove `\k` support... in this case, we use Lua's pattern instead of vim's regex.