@dmitmel 2021-09-19
> It's hard to review. So it will be take a time.

Is there anything unclear? I can split this PR into multiple ones or rewrite the history into more clean commits if you want.

> I've received the issue report (via twitter) that will make freeze the nvim when editing X mb buffer.
>
> Hm... Probably, we should remove `\k` support... in this case, we use Lua's pattern instead of vim's regex.

I respectfully disagree. While yes, rewriting `index_line` like this:

```lua
function buffer.index_line(self, linenr, line)
  -- ...
  local pos = 0
  while pos <= #line do
    local match_start, match_end = line:find(self.pattern, pos)
    if match_start and match_end then
      local word = line:sub(match_start, match_end)
      if #word > 1 then
        words[word_i] = word
        word_i = word_i + 1
      end
      pos = match_end + 1
    else
      break
    end
  end
end
```

And using `[a-zA-Z0-9_]+` as the pattern cuts the indexing time on the 10 MB file from benchmark by three times (500 ms) and reduces RAM usage by 5 MB (30 MB in total), we loose a lot of flexibility due to having to use Lua patterns. For example, your current regex is unimplementable with them because it has nested groups with repeat quantifiers, and custom regexes, in particular `\k\+`, can correctly deal with Unicode characters as can be seen from #11, this works especially well with European languages. Lua's patterns, on the other hand, have no UTF-8 support at all!

Is the file from your report minified or has very long lines, by any chance? It is known that they cause problems currently, the end of my first post contains an explanation for this. I am sure that the freezing is due to high GC load from very long slicing strings, and I am sure that it can be resolved by adding a parameter to `vim.regex.match_str` for specifying which byte to start searching from.

> So I think we should introduce `max_file_bytes` or `line_count` option.

Yes of course, I can do that. However, I still think that optimizing is worth it, since we will be able to set higher limits instead of just avoiding support of files over a limit.