@hrsh7th 2021-09-19
> Is there anything unclear? I can split this PR into multiple ones or rewrite the history into more clean commits if you want.

! Please.
Your PR isn't unclear just difficult to me 😂

OK. So we use vim.regex.