@hrsh7th 2021-09-19
BTW, I found the another problem that slows down the nvim. (it isn't relayed to cmp-buffer. Izm tackling to it now