@dmitmel 2021-11-02
By the way, I'm sorry for the quality of this PR and that it is hard to review, I just don't have much free time for work on Vim plugins these days. I actually want to help with optimization of nvim-cmp. For example, I want to try making an alternative implementation of the matcher module in C++ because it is pure algorithms and IIRC doesn't call Vim standard library functions at all. I just can't do it right now.