@dmitmel 2021-11-07
@hrsh7th I have split this PR into multiple sub-PRs, as promised. They are all based on each other's branches, here is how the history looks on my end:

![image](https://user-images.githubusercontent.com/15367354/140654033-e4eaa556-ae18-4af7-95f2-9ba076ddec22.png)

So the Github UI currently shows that e.g. #23 has commits from all previous PRs, even though each PR really has 1 or 2 commits. If the commit list is not updated when you merge previous PRs, I'll be sure to keep rebasing them one by one on top of the `main` branch of this repository.