@dmitmel 2021-11-22
Closing because I used the branch `main` for this, and having to write `origin/main` every time I want to get a diff is a bit annoying :slightly_smiling_face: